Split samples, normalization and feature extraction to different files
